pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btcxpoly"
version = "0.1.0"
description = "BTC x prediction markets research pipeline for Polymarket and Kalshi."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Codex" }
]
dependencies = [
  "duckdb>=1.2.2,<2.0.0",
  "httpx>=0.27.0,<1.0.0",
  "matplotlib>=3.9.0,<4.0.0",
  "numpy>=2.0.0,<3.0.0",
  "pandas>=2.2.0,<3.0.0",
  "pyarrow>=19.0.0,<20.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "PyYAML>=6.0.0,<7.0.0",
  "scikit-learn>=1.6.0,<2.0.0",
  "typer>=0.16.0,<1.0.0",
  "websockets>=15.0.0,<16.0.0"
]

[project.optional-dependencies]
dev = [
  "ipykernel>=6.29.0,<7.0.0",
  "jupyter>=1.1.0,<2.0.0",
  "pytest>=8.3.0,<9.0.0"
]

[project.scripts]
btcxpoly = "btcxpoly.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
